requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.11
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
openai==1.51.2
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
